# openapi/specs/b2b_dashboard.yaml
openapi: 3.1.0
info:
  title: B2B Analytics Dashboard
  description: B2B site-license analytics for org managers and MIT contract admins.
    Aggregate endpoints apply a k-anonymity floor. The contract-scoped learner-progress
    endpoint returns individual learners on the organization's own seats, with outcome
    fields consent-gated.
  version: 0.0.1
paths:
...
  /api/v1/analytics/organizations/{organization_id}/contracts/{contract_id}/learner-progress:
    get:
      tags:
      - learners
      summary: Each learner's enrollment and completion in each course run under the
        contract
      operationId: learners_progress_retrieve
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          title: Organization Id
      - name: contract_id
        in: path
        required: true
        schema:
          type: integer
          title: Contract Id
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 254
          - type: 'null'
          description: Case-insensitive match on email or name.
          title: Search
        description: Case-insensitive match on email or name.
      - name: completion_status
        in: query
        required: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/CompletionStatusFilter'
          description: Repeat for several. `unknown` selects rows with withheld outcomes.
          title: Completion Status
        description: Repeat for several. `unknown` selects rows with withheld outcomes.
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          description: Include deactivated enrollments (unenrolled, refunded).
          default: false
          title: Include Inactive
        description: Include deactivated enrollments (unenrolled, refunded).
      - name: sort
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/SortKey'
          default: full_name
      - name: descending
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Descending
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearnerProgressResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CompletionStatus:
      type: string
      enum:
      - not_started
      - in_progress
      - passed
      - certified
      title: CompletionStatus
      description: '``passed`` without ``certified`` is normal: certificates are issued
        on a

        schedule after grading, and audit-mode enrollments never certify.'
    CompletionStatusFilter:
      type: string
      enum:
      - not_started
      - in_progress
      - passed
      - certified
      - unknown
      title: CompletionStatusFilter
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LearnerProgress:
      properties:
        learner_id:
          type: string
          title: Learner Id
          description: The learner's account ID. It stays the same if their email
            address changes.
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
          description: The learner's email address. It may differ from the one they
            enrolled with.
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
          description: The learner's name, if they've provided one.
        courserun_readable_id:
          type: string
          title: Courserun Readable Id
          description: The course run's ID, e.g. course-v1:MITxT+14.310x+2T2026.
        courserun_title:
          type: string
          title: Courserun Title
          description: The course's title.
        courserun_start_on:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Courserun Start On
          description: When the course run starts. Empty if no start date is set.
        courserun_end_on:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Courserun End On
          description: When the course run ends. Empty for self-paced courses.
        enrolled_on:
          type: string
          format: date-time
          title: Enrolled On
          description: When the learner enrolled in this course run.
        enrollment_is_active:
          type: boolean
          title: Enrollment Is Active
          description: Whether the learner is still enrolled. False if the enrollment
            was deactivated, for example after unenrolling.
        enrollment_mode:
          anyOf:
          - type: string
          - type: 'null'
          title: Enrollment Mode
          description: The enrollment track, for example verified or audit. Audit
            enrollments don't earn certificates.
        outcomes_shared:
          type: boolean
          title: Outcomes Shared
          description: Whether the learner has agreed to share their progress. If
            not, their status, grades, certificate and activity are hidden.
        completion_status:
          anyOf:
          - $ref: '#/components/schemas/CompletionStatus'
          - type: 'null'
          description: 'Where the learner is in the course: not started, in progress,
            passed or certified. Hidden if the learner hasn''t agreed to share their
            progress.'
        is_passing:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Passing
          description: Whether the learner currently has a passing grade. Empty if
            no grade has been calculated yet. Hidden if the learner hasn't agreed
            to share their progress.
        grade:
          anyOf:
          - type: number
          - type: 'null'
          title: Grade
          description: The learner's current grade, from 0 to 1. Hidden if the learner
            hasn't agreed to share their progress.
        letter_grade:
          anyOf:
          - type: string
          - type: 'null'
          title: Letter Grade
          description: The learner's letter grade, when the course assigns one. Hidden
            if the learner hasn't agreed to share their progress.
        certificate_issued_on:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Certificate Issued On
          description: When the learner's certificate was issued. Empty if they don't
            have one. Hidden if the learner hasn't agreed to share their progress.
        certificate_is_revoked:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Certificate Is Revoked
          description: Whether the learner's certificate was revoked. A revoked certificate
            doesn't count toward completion. Empty if they don't have one. Hidden
            if the learner hasn't agreed to share their progress.
        last_active_on:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Last Active On
          description: The last day the learner did anything in the course. Not available
            yet, so always empty for now.
      type: object
      required:
      - learner_id
      - email
      - full_name
      - courserun_readable_id
      - courserun_title
      - courserun_start_on
      - courserun_end_on
      - enrolled_on
      - enrollment_is_active
      - enrollment_mode
      - outcomes_shared
      - completion_status
      - is_passing
      - grade
      - letter_grade
      - certificate_issued_on
      - certificate_is_revoked
      - last_active_on
      title: LearnerProgress
      description: One learner's enrollment in one course run under the contract.
    LearnerProgressResponse:
      properties:
        organization_id:
          type: string
          title: Organization Id
          description: The organization's ID.
        as_of:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
          description: When the data was last updated. Empty before the first update.
        total_count:
          type: integer
          title: Total Count
          description: Matching enrollments across all pages, not just this one.
        outcomes_withheld_count:
          type: integer
          title: Outcomes Withheld Count
          description: How many of those enrollments have progress hidden because
            the learner hasn't agreed to share it.
        data:
          items:
            $ref: '#/components/schemas/LearnerProgress'
          type: array
          title: Data
          description: This page of enrollments.
      type: object
      required:
      - organization_id
      - as_of
      - total_count
      - outcomes_withheld_count
      - data
      title: LearnerProgressResponse
      description: 'The org envelope (``organization_id``, ``as_of``, ``total_count``,

        ``data``) plus ``outcomes_withheld_count``, so a client can show how many

        rows carry withheld outcomes without paging through all of them.'
...
    SortKey:
      type: string
      enum:
      - full_name
      - email
      - enrolled_on
      - courserun_readable_id
      title: SortKey
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError